rename string_work into serial_work
