--- task_02/in_1.txt
@@ -1,1 +1,0 @@
-foHKDwgLYHpWNRTGvyQdECQoqIgRWiPPitZyTXyNQExcdWGwxHWWfoxHRrILhtfozdCFmGeVgtRBnZzXfoceRwcrsnrDUCaXNGMQ
