--- task_02/in_2.txt
@@ -1,1 +1,0 @@
-vibesignbikefootbearsignneckneckwinedoorpagebeerkiteweekroadbikevibebeardoorsignkitebearwinebearsign
